Fix Slack notifications by using shared 'slack-notification-token'
Some of our Slack notifications have been broken since at least
February 2021:
https://github.com/alphagov/govuk-puppet/pull/10922#issuecomment-772484768

Testing the `SECOND_LINE_SLACK_AUTH_TOKEN` token against
https://slack.com/api/auth.test, it doesn't appear to be valid.
But the `slack-notification-token` (added at some point in the
past, in the Jenkins UI: https://github.com/jenkinsci/slack-plugin/issues/270#issuecomment-276176109)
DOES seem to work.

We can see in the UI for any of the jobs in this commit, that the
'Integration Token' is set, and the 'Integration Token Credential
ID' is not set. Clicking 'Test Connection', we see a client error.
If we clear the 'Integration Token' field, and instead set the
'Integration Token Credential ID' to 'slack-notification-token',
the 'Test Connection' succeeds, and we see the test message come
through to Slack successfully.

This commit therefore removes the reliance on API keys stored in
govuk-secrets, and instead makes reference to a shared secret
stored in Jenkins itself. This should fix Slack notifications for
4 different jobs.

![](https://user-images.githubusercontent.com/5111927/180235908-3bf5f3bb-387b-45c3-a736-e1ef8fb6512f.png)

Trello: https://trello.com/c/aGpBjdHG/2899-fix-slack-notifications-for-failed-jenkins-jobs3