Update forms hostname for signon db sync
This adds the forms service to the list of transformations that are done to signon as part of it's sync from production to staging. This is required so that the forms service can utilise the GOV.UK staging signon service from its staging environment.

An alternative approach could be that we create a staging environment in Signon production, as it's not necessary that the staging app of forms uses the staging app of signon. However this would go against the existing conventions used with other 3rd party applications and thus seems something to only explore if we experience problems.