Remove ASSET_HOST for Whitehall
This set originally to deal with CSV asset preview pages that were
hosted on the assets.publishing.service.gov.uk domain, but rendered by
Whitehall. A proxy is set up to now also forward Whitehall asset (css,
js etc) requests on the asset domain. This means that we can use
relative paths for assets as they are now able to be served on the same
domain.