Add Referrer-Policy header
The Referrer-Policy HTTP header controls how much referrer information (sent with the Referer header) should be included with requests.

This commit sets the header to "strict-origin-when-cross-origin" to aid in security.

The header will not allow any information to be sent when a scheme downgrade happens (the user is navigating from HTTPS to HTTP).