Remove AWS related heira config for Ask Service (production)
There was a critical alert in Production due to the Ask data export failing to export to S3. [Logs here](https://deploy.blue.production.govuk.digital/job/govuk-ask-export/289/console)

Given that we do not own the bucket and suspect that it isn't in use, we have decided to remove exporting the Ask Service data to S3. The data will continue to be exported to Google Drive.