Use s5cmd to drastically speed up sync_mirror.
Unlike `s3cmd sync` (and `aws s3 sync`), s5cmd performs operations in
parallel and is typically an order of magnitude or more faster than the
former. https://github.com/peak/s5cmd/#benchmarks

It also uses less resources.

Considered using Puppet to install s5cmd, but it's just not worth it.
Installing it via this script is just as maintainable (if not more) and
way simpler.