Increase proxy buffer size in nginx for all machines
A change to the way CSS is loaded sees the `preload` header increase in size
with the additional CSS files. This causes nginx to throw the error `upstream
sent too big header while reading response header from upstream`.

The error wasn't present on the draft stack because that has had the proxy
buffer size increased in #11715[1] - so this pull request makes that change to
the other stacks as well to prevent this error occurring in the future.

[1]: https://github.com/alphagov/govuk-puppet/pull/11715