Use environment specific assets hostname alias
This configures assets so it can receive requests from asset hostnames
in it's respective environment (e.g.
assets.integration.publishing.service.gov.uk)

It was unusual for this to be configured, as previous, where only the
production hostname was set.

The motivation for making this change is to remove an inconsistency in
our CDN config where for assets we have Fastly reset the request host.
In our other CDN configs the host header is not changed.