Remove 'Launch_VMs' Jenkins tasks
These haven't been updated in around 5 years, and look as though
they were only used for provisioning virtual machines in Carrenza.
They all seem to rely on the govuk-provisioning repo, which is now
archived. And finally, they don't seem to be being pulled in as
jobs into any of the Jenkins environments, so are ultimately not
being used.