Move proxy buffer headers to proxy nginx location
Previously the proxy buffer headers were set in the root location (/) of
the nginx declaration and not the declaration used to proxy a request.
This meant that these headers weren't having any effect and hadn't
resolved the problem.

Moving these headers to the proxy declaration means they have an effect.