Remove failure alert from mirrorer script
This prevents the mirrorer script for alerting that the script has
failing. This is an unnecessary alert, as we already have an alarm that
is raised when the script hasn't reported to have successfully completed
in the last 24hrs. The failure alert is noise as the script typically
re-runs, however takes another ~10hrs, therefore we're stuck with a
failure alert that is unactionable.