Run PagerDuty drill every day
We usually only run the drill on Wednesday mornings, when there is
a handover of support from one team to the next. Over the
Christmas period, support changes on a more granular basis; there
are different first responders between 9:30-17:30 and 17:30-9:30,
with rotations happening twice per day.

This change will be reverted on or around the 4th of January.

Note that the weekday must be set to absent, otherwise it will
not automatically reset to '*', it would just stay as Wednesday.
See https://puppet.com/docs/puppet/5.5/types/cron.html