Prevent "high memory for <app-name>-procofile-worker" warning
Platform Health removed the parent app memory alerts in
https://trello.com/c/nfzw7FvL/2458-remove-app-memory-alerts-2.
We forgot to clear up the app's workers while we were there.

It was agreed that these kinds of alerts aren't actionable and
are just added noise. What happens when we have high memory for
a worker? It is rarely a useful indicator of anything.

This commit ensures that the alert/warning is removed. We can
then clear up the associated code.

Trello: https://trello.com/c/12p4D0if/3068-remove-high-memory-for-appname-procfile-worker-alert