Bump CPU thresholds for content-store in Staging
The content-store app is handling more traffic on Staging than
Production, because the GoReplay traffic replay from Production
is replaying traffic to both the EC2 and EKS platforms, but both
platforms are using the EC2 content store.

We're therefore regularly seeing CPU warnings for content-store
on Staging. They seem to spike to no higher than around 220%, and
none of the requests seem to error, so Content Store is coping
with the demand.

We should therefore be able to set a comfortable limit of say
250%, but we only want to apply this to Staging, where we're in
this weird world of double platforms. Therefore I've made the
thresholds a configurable field that is set in YAML.

I did consider duplicating the `staging/content_store.yaml` into
a `staging/draft_content_store.yaml` file, but we've only seen the
alert on the non-draft content store, which makes sense as the
bulk of the traffic is replayed from (public) production content
store requests.

Trello: https://trello.com/c/T2aatFyl/1516-high-cpu-usage-for-content-store-app-on-staging