Remove 'Smokey Loop' code
We currently have two ways of running Smokey:
1) Via the Jenkins job (e.g. https://deploy.blue.staging.govuk.digital/job/Smokey/)
2) Via the 'Smokey Loop', which runs continuously on the
   `monitoring` machine in each environment.

Having to maintain the two different approaches is a technical
difficulty, made all the more difficult by newly provisioned
`monitoring` instances no longer installing `google-chrome-stable`.
Puppet fails with:

```
The following packages have unmet dependencies:
 google-chrome-stable : Depends: libdbus-1-3 (>= 1.9.14) but 1.6.18-0ubuntu4.5+esm3 is to be installed
                        Depends: libgbm1 (>= 17.1.0~rc2) but it is not going to be installed
                        Depends: libxkbcommon0 (>= 0.5.0) but 0.4.1-0ubuntu1.1 is to be installed
                        Recommends: libu2f-udev but it is not installable
                        Recommends: libvulkan1 but it is not installable
```

Even if we were to find a workaround for this, part of the reason
we've been provisioning new `monitoring` instances is that the
existing instances had been struggling with CPU load. Having to
run Chrome and a continuous Smokey loop is only going to
exacerbate those issues.

Therefore, we plan to remove the existing 'Smokey Loop' code
entirely, and instead use the more reliable 'Jenkins job' approach
to running Smokey tests. We can configure this job to run in a
regular cron pattern, ensuring we still retain the 'probe test'
coverage that was once provided by the Smokey Loop.

What we'll lose is the Icinga alerting if Smokey fails, but this
will be replaced by a Slack notification (already supported by
the Jenkins approach).