Send procfile worker stdout to logit via filebeat
This configures govuk::procfile::worker to utilise filebeat to ship log
files. The motivation for adding this is that GOV.UK apps that use
Sidekiq >= 6 are not having their logs shipped to logit.

The most common usage of govuk::procfile::worker is for Sidekiq and
prior to version 6 GOV.UK wrote Sidekiq logs to a file
(log/sidekiq.json). With Sidekiq 6 there is no longer the ability to
write to a logfile [1]. While there have previously been stdout and
stderr log files created as part of a govuk::procfile::worker process
these have not been configured to use filebeat.

While Sidekiq is the most common usage of govuk::procfile::worker it is
not the only one - it is used for running other background tasks like
RabbitMQ consumers - and thus some compromises were made so it wasn't
Sidekiq specific. This mostly impacts the tagging of logs which will now
use the process name from the procfile (generally "worker") rather than
a tag of "sidekiq".

The expectation is that GOV.UK Sidekiq logging will be in JSON so this
is the default in the class. For other uses of this class there's a good
chance that this will need to be set to false (see next commit).

In experimenting with this I investigated whether to ship stderr logs as
well as stdout and became concerned that this might create a yak shave.
From my brief testing there seemed to be a lot of junk in stderr
(search-api logged a Ruby warning every 4 seconds 😭) and there seemed
to be some duplications (presumably a consequence of our stdout/stderr
munging in GovukLogging [2]). As these logs weren't previously shipped
adding this seemed to be more adding a new feature that fixing something
lost so I stepped away from it.

[1]: https://www.mikeperham.com/2019/09/03/welcome-to-sidekiq-6.0/
[2]: https://github.com/alphagov/govuk_app_config/blob/13083f8f8563c4703c14fe8175ca35646ec64442/lib/govuk_app_config/govuk_logging.rb#L7-L26