Configure non JSON logging for govuk::procfile::worker usage
In the previous commit we changed govuk::procfile::worker to ship
procfile worker stdout logs to logit via filebeat.
govuk::procfile::worker expects the stdout log file to be structured
JSON logging. This commit changes the usage of this class which aren't
expected to output JSON.

To verify these I checked the log files on an integration box, and for
the ones that had data I couldn't see JSON formatting. If I missed any
and they log JSON it won't be a big deal it'll just mean a JSON string
is in the log - this is preferable to getting it wrong the other way
where errors would be reported for invalid JSON.