Run a Redis 6 container instance in CI
On our Jenkins CI machines we run Redis 2.8.4 😭, this finally caught up
to us with tests for GOV.UK apps that use Sidekiq 6 and makes use of
newer Redis features.

This applies the same approach we've used to get around ancient versions
of PostgreSQL, MongoDB and MySQL and sets up a containerised version of
Redis so our apps can use an equivalent to version as the production
runtime.

This sets-up a Redis 6 container that runs on the CI machines (we are
using Redis 6 via ElastiCache in production). Rails applications can use
this by setting up an env var e.g:

```
  govuk.setEnvar("REDIS_URL", "redis://127.0.0.1:63796")
```

I set the port number to be 63796 to reflect 6379 as the regular Redis
port and 6 to signify the version.

An example of this running successfully is: https://ci.integration.publishing.service.gov.uk/job/travel-advice-publisher/job/sidekiq-6-kevin/6/