Remove `content_data_api_re_run` job from Staging & Integration
The job hasn't succeeded for almost 6 months in Staging, and for
an entire year in Integration! It is not worth fixing as this job
only exists to fix a race condition where Production was a day or
two behind (see #10437).

We'll keep the hack in Production, and allow Staging & Integration
to receive the most up-to-date results through the nightly env
sync. We shouldn't accept having perpetually failing jobs in any
environment.

![Job on Staging](https://user-images.githubusercontent.com/5111927/215437810-c66de700-df0e-43a0-9eb0-f4470d6a8a1e.png)
![Job on Integration](https://user-images.githubusercontent.com/5111927/215437813-6dd8100f-3863-47cb-975b-60d275f0be76.png)