Prevent govuk_env_sync from waiting on user input.
When govuk_env_sync runs with a tty, for example on a manual run under
screen or tmux or in an ssh session with `-t`, `psql` will spawn a pager
(like `less`) and wait indefinitely for user input before continuing if
a command produces more than a certain volume of output. This leads to
one-off runs getting stuck.

Prevent this by disabling the `pager` output option in `psql`. See
`\pset` under https://www.postgresql.org/docs/9.0/app-psql.html (it's
the same in later versions).

Normally I'd do a lot more cleanup for something like this, but the
script is in pretty bad shape to start with and I don't have the time
right now.