Remove Data_Sync_Complete cron triggers on Jenkins.
Some of the functionality of this job is no longer needed; the rest is
handled in https://github.com/alphagov/govuk-helm-charts/pull/1118.

Holding off on deleting the jobs entirely for a couple of days just so
that it's easy to run them manually if it turns out I've missed
something.