CSV previews should work with modern urls
We are migrating away from legacy url path in whitehall assets. Router and nginx config has a special list of urls that are directed to frontend instead of whitehall. The change is to include the new url in the configuration.

After adding the new url to whitehall_uploaded_assets_routes, it will be used by both assets-origin and draft-assets.