Get rid of old Vagrant credentials.
The Vagrant environment is long-dead (and only ever would have been run
locally) but it's not a good look to have any kind of private key or
password checked into a public repo (even if it's only used for
testing), so let's clean it up.

I tried just deleting these files, but turns out the pre-merge testsuite
uses them so I've just replaced the secret-looking values with dummy
values.

Addresses #12155, with credit to @annaobarz for bringing this to our
attention.